Version_1.1/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
PyMuPDF>=1.24.0
llama-cpp-python>=0.2.90
huggingface-hub>=0.24.0
numpy<2.0.0
pydantic>=2.0.0
tiktoken>=0.7.0
